different clean up and refactoring
